Divided the script into multiple files for better organization and maintainability.
