+++ src/index.html
@@ -195,7 +195,7 @@
 
 </div>
 
-<script src="script.js"></script>
+<script type="module" src="js/main.js"></script>
 
 </body>
 </html>
