+++ src/js/data.js
@@ -1,0 +1,94 @@
+export const adjectives = [
+    "Radioactive", "Invisible", "Sticky", "Ancient", "Digital", "Angry", "Wet", "Frozen", "Gigantic", "Tiny", 
+    "Floating", "Explosive", "Silent", "Glowing", "Magnetic", "Edible", "Venomous", "Indestructible", "Haunted", 
+    "Furry", "Transparent", "Clockwork", "Liquid", "Infinite", "Broken", "Sentient", "Paper", "Electric", "Smelly", 
+    "Expensive", "Useless", "Dangerous", "Holy", "Cursed", "Heavy", "Feather-light", "Solar-powered", "Underground", 
+    "Flying", "Inflatable", "Recycled", "Origami", "Concrete", "Neon", "Musical", "Intelligent", "Telepathic", 
+    "Unstoppable", "Microscopic", "Biodegradable", "Hyper-intelligent", "Self-replicating", "Emotionally unstable", 
+    "Glitchy", "Slightly radioactive", "Invisibly vibrating", "Made of meat", "Time-traveling", "Possessed", 
+    "Bureaucratic", "Philosophical", "Passive-aggressive", "Overly enthusiastic", "Non-Euclidean", "Single-use", 
+    "Screaming", "Hypnotic", "Existentially Dreadful", "Slightly Damp", "Suspiciously Moist"
+];
+
+export const nouns = [
+    "Brick", "Toaster", "Smartphone", "Sock", "Banana", "Chainsaw", "Cloud", "Piano", "Cactus", "Skeleton", 
+    "Shopping Cart", "Ladder", "Toothbrush", "Paperclip", "Bucket", "Cat", "Dinosaur", "Car Trunk", "Tire", 
+    "Coffee Mug", "Alarm Clock", "Matchbox", "Shoe", "Umbrella", "Mirror", "Hammer", "Bicycle", "Toilet", 
+    "Telescope", "Sandwich", "Refrigerator", "Drone", "Statue", "Treehouse", "Submarine", "Helicopter", 
+    "Lightbulb", "Pillow", "Guitar", "Spoon", "Key", "Wallet", "Backpack", "Chair", "Table", "Door", "Window", 
+    "Book", "Pen", "Computer", "Egg", "Ocean", "Mountain", "Volcano", "Tax Return", "Voodoo Doll", 
+    "Quantum Computer", "Used Tissue", "Bag of Holding", "Black Hole", "Rubber Duck", "Grandmother's Vase", 
+    "Nuclear Launch Button", "Hamster Wheel", "Traffic Cone", "Ouija Board", "Sourdough Starter", "NFT", 
+    "Influencer", "Politician", "Clown Wig", "Mayonnaise Jar", "Unexploded Ordnance", "Love Letter", "Human Soul"
+];
+
+export const primingAngles = {
+    functional: [
+        "Disaster / Survival", "Weaponized / Aggressive", "Repair / Construction", 
+        "Musical / Artistic", "Legal / Bureaucratic", "Crime / Theft", 
+        "Luxury / Decoration", "Low-Effort / Lazy", "High-Tech Upgrade", "Zero-G Environment",
+        "Zombie Apocalypse", "Espionage / Spycraft", "Medical Emergency"
+    ],
+    hypothetical: [
+        "Sci-Fi / Futuristic", "Prehistoric / Primitive", "Fairy Tale / Magical", 
+        "Historical Event", "Anthropological / Cultural", "Religious / Ritual", 
+        "Political / Governing", "Philosophical Meaning", "Post-Apocalyptic", "Children's Story",
+        "Alien Contact", "Parallel Dimension", "Cyberpunk Dystopia"
+    ],
+    opposite: [
+        "Financial Ruin", "Social Embarrassment", "Legal Consequence", 
+        "Physical Pain", "Ecological Disaster", "Extreme Inconvenience", 
+        "Instant Failure", "Toxic / Poisonous", "Public Outrage", "Utter Uselessness",
+        "Existential Crisis", "Public Humiliation", "Accidental Treason"
+    ]
+};
+
+export const functionalTemplates = [
+    "How would you use {a/an} {adj_word} {noun_word} to escape a prison?", 
+    "Use {a/an} {adj_word} {noun_word} to fix a broken car.", 
+    "How would a billionaire use {a/an} {adj_word} {noun_word}?", 
+    "Invent a sport played with {a/an} {adj_word} {noun_word}.", 
+    "How would you use {a/an} {adj_word} {noun_word} for self-defense?", 
+    "Sell {a/an} {adj_word} {noun_word} on a TV infomercial.",
+    "You are a supervillain. How does {a/an} {adj_word} {noun_word} fit into your master plan?",
+    "Design a trap for a ghost using {a/an} {adj_word} {noun_word}.",
+    "How would you use {a/an} {adj_word} {noun_word} to win a rap battle?",
+    "You are stranded on Mars. How does {a/an} {adj_word} {noun_word} save your life?"
+];
+
+export const hypotheticalTemplates = [
+    "You are a ghost. How do you haunt someone with {a/an} {adj_word} {noun_word}?", 
+    "It is the year 3000. Archaeologists find {a/an} {adj_word} {noun_word}. What do they think it is?", 
+    "Write a news headline about {a/an} {adj_word} {noun_word}.", 
+    "If {a/an} {adj_word} {noun_word} became the president, what is its first law?", 
+    "Describe a religion that worships {a/an} {adj_word} {noun_word}.", 
+    "A wizard turns you into {a/an} {adj_word} {noun_word}. What do you do?",
+    "In a world where {a/an} {adj_word} {noun_word} is currency, how do you become rich?",
+    "{a/an} {adj_word} {noun_word} is now the only evidence in a murder trial. What happened?",
+    "Aliens interpret {a/an} {adj_word} {noun_word} as a declaration of war. Why?"
+];
+
+export const oppositeTemplates = [
+    "How to accidentally kill someone with {a/an} {adj_word} {noun_word}?", 
+    "Why is {a/an} {adj_word} {noun_word} the worst gift for a toddler?", 
+    "Describe the most illegal use for {a/an} {adj_word} {noun_word}.", 
+    "Why would {a/an} {adj_word} {noun_word} ruin a wedding?", 
+    "Ways {a/an} {adj_word} {noun_word} could destroy the world.", 
+    "What is the most boring thing to do with {a/an} {adj_word} {noun_word}?",
+    "How to lose all your friends using only {a/an} {adj_word} {noun_word}.",
+    "Why is {a/an} {adj_word} {noun_word} banned in 17 countries?",
+    "Describe the most awkward first date involving {a/an} {adj_word} {noun_word}."
+];
+
+export const wildcards = [
+    "⚡ New Rule: No using your hands!",
+    "⚡ New Rule: Must involve water!",
+    "⚡ New Rule: Cost must be under $10!",
+    "⚡ New Rule: Must be silent!",
+    "⚡ New Rule: Must be edible!",
+    "⚡ New Rule: Must be invisible!",
+    "⚡ New Rule: Must be heavy!",
+    "⚡ New Rule: Must be tiny!",
+    "⚡ New Rule: Must be made of paper!",
+    "⚡ New Rule: Must be dangerous!"
+];
